Intermediate commit: Start to separate JDK server and handlers into a transport extension
