+++ client/src/environments/environment.localhttps.ts
@@ -20,7 +20,6 @@
 
 export const environment = {
     production: false,
-    apiUrl: 'https://localhost:3000/_/',
-    apiPostUrl: 'https://localhost:8080'
-
+    apiUrl: 'https://dataportabilitydemo2.westeurope.cloudapp.azure.com:3000/_/',
+    apiPostUrl: 'https://dataportabilitydemo2.westeurope.cloudapp.azure.com:8080'
 };
