+++ distributions/demo-google-deployment/api/build.gradle
@@ -53,6 +53,7 @@ repositories {
 
 dependencies {
     compile project(':portability-api')
+    compile project(':extensions:transport:portability-transport-jdk')
     compile project(':extensions:config:portability-config-yaml')
     // TODO: depend on these based on list in flag values.
     compile project(':extensions:auth:portability-auth-microsoft')
