+++ distributions/demo-server/build.gradle
@@ -59,6 +59,7 @@ dependencies {
     compile project(':portability-bootstrap-vm')
 
     compile project(':portability-api')
+    compile project(':extensions:transport:portability-transport-jdk')
     compile project(':portability-transfer')
 
     compile project(':extensions:config:portability-config-yaml')
