+++ distributions/demo-server/src/main/resources/config/api.yaml
@@ -1,2 +1,2 @@
-baseUrl: https://localhost:3000
-baseApiUrl: https://localhost:8080
+baseUrl: https://dataportabilitydemo2.westeurope.cloudapp.azure.com:3000
+baseApiUrl: https://dataportabilitydemo2.westeurope.cloudapp.azure.com:8080
