+++ extensions/transport/portability-transport-jdk/src/main/java/org/dataportabilityproject/transport/jdk/http/ImportSetupHandler.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.dataportabilityproject.api.reference;
+package org.dataportabilityproject.transport.jdk.http;
 
 import com.google.inject.Inject;
 import com.google.inject.name.Named;
