+++ extensions/transport/portability-transport-jdk/src/main/java/org/dataportabilityproject/transport/jdk/http/JsonKeys.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.dataportabilityproject.api.reference;
+package org.dataportabilityproject.transport.jdk.http;
 
 /** Keys for returned Json data. */
 public class JsonKeys {
