+++ extensions/transport/portability-transport-jdk/src/main/java/org/dataportabilityproject/transport/jdk/http/ListDataTypesHandler.java
@@ -13,64 +13,59 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.dataportabilityproject.api.reference;
-
-import static com.google.common.net.HttpHeaders.CONTENT_TYPE;
+package org.dataportabilityproject.transport.jdk.http;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.base.Preconditions;
-import com.google.common.base.Strings;
 import com.google.inject.Inject;
 import com.sun.net.httpserver.Headers;
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-import org.dataportabilityproject.api.action.listservices.ListServicesAction;
-import org.dataportabilityproject.api.action.listservices.ListServicesActionRequest;
-import org.dataportabilityproject.api.action.listservices.ListServicesActionResponse;
 import org.dataportabilityproject.api.launcher.TypeManager;
-import org.dataportabilityproject.api.reference.ReferenceApiUtils.HttpMethods;
-import org.dataportabilityproject.types.client.transfer.ListServicesResponse;
+import org.dataportabilityproject.api.action.listdatatypes.ListDataTypesAction;
+import org.dataportabilityproject.api.action.listdatatypes.ListDataTypesActionRequest;
+import org.dataportabilityproject.api.action.listdatatypes.ListDataTypesActionResponse;
+import org.dataportabilityproject.transport.jdk.http.ReferenceApiUtils.HttpMethods;
+import org.dataportabilityproject.types.client.transfer.ListDataTypesResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-/** HttpHandler for the {@link ListServicesAction}. */
-final class ListServicesHandler implements HttpHandler {
+import java.io.IOException;
+import java.nio.charset.StandardCharsets;
 
-  public static final String PATH = "/_/listServices";
-  private static final Logger logger = LoggerFactory.getLogger(ListServicesHandler.class);
-  private final ListServicesAction listServicesAction;
+import static com.google.common.net.HttpHeaders.CONTENT_TYPE;
+
+/** HttpHandler for the {@link ListDataTypesAction}. */
+final class ListDataTypesHandler implements HttpHandler {
+
+  public static final String PATH = "/_/listDataTypes";
+  private static final Logger logger = LoggerFactory.getLogger(ListDataTypesHandler.class);
+  private final ListDataTypesAction listDataTypesAction;
   private final ObjectMapper objectMapper;
 
   @Inject
-  ListServicesHandler(ListServicesAction listServicesAction, TypeManager typeManager) {
-    this.listServicesAction = listServicesAction;
+  ListDataTypesHandler(ListDataTypesAction listServicesAction, TypeManager typeManager) {
+    this.listDataTypesAction = listServicesAction;
     this.objectMapper = typeManager.getMapper();
   }
 
-  /** Services the {@link ListServicesAction} via the {@link HttpExchange}. */
+  /** Services the {@link ListDataTypesAction} via the {@link HttpExchange}. */
   @Override
   public void handle(HttpExchange exchange) throws IOException {
     Preconditions.checkArgument(ReferenceApiUtils.validateRequest(exchange, HttpMethods.GET, PATH));
     logger.debug("received request: {}", exchange.getRequestURI());
 
-    String transferDataType = ReferenceApiUtils.getRequestParams(exchange).get(JsonKeys.DATA_TYPE);
-    Preconditions.checkArgument(!Strings.isNullOrEmpty(transferDataType), "Missing data type");
-
-    ListServicesActionRequest actionRequest = new ListServicesActionRequest(transferDataType);
-    ListServicesActionResponse actionResponse = listServicesAction.handle(actionRequest);
+    ListDataTypesActionResponse actionResponse =
+        listDataTypesAction.handle(new ListDataTypesActionRequest());
 
     if (actionResponse.getErrorMsg() != null) {
       logger.warn("Error during action: {}", actionResponse.getErrorMsg());
-      handleError(exchange, transferDataType);
+      handleError(exchange);
       return;
     }
 
-    String[] importServices = actionResponse.getImportServices().toArray(new String[0]);
-    String[] exportServices = actionResponse.getExportServices().toArray(new String[0]);
-    ListServicesResponse response =
-        new ListServicesResponse(transferDataType, exportServices, importServices);
+    String[] dataTypes = actionResponse.getTransferDataTypes().toArray(new String[0]);
+    ListDataTypesResponse response = new ListDataTypesResponse(dataTypes);
 
     // Set response as type json
     Headers headers = exchange.getResponseHeaders();
@@ -82,9 +77,9 @@ final class ListServicesHandler implements HttpHandler {
   }
 
   /** Handles error response. TODO: Determine whether to return user facing error message here. */
-  public void handleError(HttpExchange exchange, String transferDataType) throws IOException {
+  public void handleError(HttpExchange exchange) throws IOException {
     String[] empty = new String[] {};
-    ListServicesResponse response = new ListServicesResponse(transferDataType, empty, empty);
+    ListDataTypesResponse response = new ListDataTypesResponse(empty);
     // Mark the response as type Json and send
     exchange
         .getResponseHeaders()
