+++ extensions/transport/portability-transport-jdk/src/main/java/org/dataportabilityproject/transport/jdk/http/Oauth2CallbackHandler.java
@@ -13,25 +13,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.dataportabilityproject.api.reference;
+package org.dataportabilityproject.transport.jdk.http;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.google.api.client.auth.oauth2.AuthorizationCodeResponseUrl;
 import com.google.common.base.Preconditions;
 import com.google.common.base.Strings;
 import com.google.common.io.BaseEncoding;
+import com.google.common.net.HttpHeaders;
 import com.google.inject.Inject;
 import com.google.inject.name.Named;
 import com.sun.net.httpserver.Headers;
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
 import java.io.IOException;
-import java.nio.charset.StandardCharsets;
+import java.net.HttpCookie;
+import java.util.Map;
 import java.util.UUID;
 import javax.crypto.SecretKey;
-import org.apache.http.HttpHeaders;
 import org.dataportabilityproject.api.launcher.TypeManager;
-import org.dataportabilityproject.api.reference.ReferenceApiUtils.FrontendConstantUrls;
-import org.dataportabilityproject.api.reference.ReferenceApiUtils.HttpMethods;
+import org.dataportabilityproject.security.DecrypterFactory;
 import org.dataportabilityproject.security.EncrypterFactory;
 import org.dataportabilityproject.security.SymmetricKeyGenerator;
 import org.dataportabilityproject.spi.cloud.storage.JobStore;
@@ -39,23 +40,21 @@ import org.dataportabilityproject.spi.cloud.types.PortabilityJob;
 import org.dataportabilityproject.spi.api.auth.AuthDataGenerator;
 import org.dataportabilityproject.spi.api.auth.AuthServiceProviderRegistry;
 import org.dataportabilityproject.spi.api.auth.AuthServiceProviderRegistry.AuthMode;
-import org.dataportabilityproject.types.client.transfer.DataTransferResponse;
-import org.dataportabilityproject.types.client.transfer.DataTransferResponse.Status;
-import org.dataportabilityproject.types.client.transfer.SimpleLoginRequest;
 import org.dataportabilityproject.types.transfer.auth.AuthData;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- * HttpHandler for SimpleLoginSubmit authorization flow. Redirects client request to: - the next
+ * HttpHandler for callbacks from Oauth2 authorization flow. Redirects client request to: - the next
  * authorization (if this is after the source service auth) or - the copy page (if this is after the
  * destination service auth)
  */
-final class SimpleLoginSubmitHandler implements HttpHandler {
+final class Oauth2CallbackHandler implements HttpHandler {
 
-  public static final String PATH = "/_/simpleLoginSubmit";
-  private static final Logger logger = LoggerFactory.getLogger(SimpleLoginSubmitHandler.class);
-
+  public static final String PATH = "/callback/";
+  // TODO: obtain from flags
+  private static final boolean IS_LOCAL = true;
+  private final Logger logger = LoggerFactory.getLogger(Oauth2CallbackHandler.class);
   private final AuthServiceProviderRegistry registry;
   private final JobStore store;
   private final SymmetricKeyGenerator symmetricKeyGenerator;
@@ -64,7 +63,7 @@ final class SimpleLoginSubmitHandler implements HttpHandler {
   private final String baseApiUrl;
 
   @Inject
-  SimpleLoginSubmitHandler(
+  Oauth2CallbackHandler(
       JobStore store,
       AuthServiceProviderRegistry registry,
       SymmetricKeyGenerator symmetricKeyGenerator,
@@ -79,102 +78,123 @@ final class SimpleLoginSubmitHandler implements HttpHandler {
     this.baseApiUrl = baseApiUrl;
   }
 
+  @Override
   public void handle(HttpExchange exchange) throws IOException {
+    // Add .* to resource path as this path will be of the form /callback/SERVICEPROVIDER
     Preconditions.checkArgument(
-        ReferenceApiUtils.validateRequest(exchange, HttpMethods.POST, PATH));
+        ReferenceApiUtils.validateRequest(exchange, ReferenceApiUtils.HttpMethods.GET, PATH + ".*"));
     logger.debug("received request: {}", exchange.getRequestURI());
 
-    DataTransferResponse response = handleExchange(exchange);
-
-    logger.debug("redirecting to: {}", response.getNextUrl());
-    // Mark the response as type Json and send
-    exchange
-        .getResponseHeaders()
-        .set(
-            HttpHeaders.CONTENT_TYPE, "application/json; charset=" + StandardCharsets.UTF_8.name());
-    exchange.sendResponseHeaders(200, 0);
-
-    objectMapper.writeValue(exchange.getResponseBody(), response);
+    String redirect = handleExchange(exchange);
+    logger.debug("redirecting to {}", redirect);
+    exchange.getResponseHeaders().set(HttpHeaders.LOCATION, redirect);
+    exchange.sendResponseHeaders(303, -1);
   }
 
-  DataTransferResponse handleExchange(HttpExchange exchange) throws IOException {
+  private String handleExchange(HttpExchange exchange) throws IOException {
+    String redirect = "/error";
 
-    Headers requestHeaders = exchange.getRequestHeaders();
     try {
-      SimpleLoginRequest request =
-          objectMapper.readValue(exchange.getRequestBody(), SimpleLoginRequest.class);
+      Headers requestHeaders = exchange.getRequestHeaders();
 
-      String encodedIdCookie = ReferenceApiUtils.getCookie(requestHeaders, JsonKeys.ID_COOKIE_KEY);
+      String requestURL =
+          ReferenceApiUtils.createURL(
+              requestHeaders.getFirst(HttpHeaders.HOST),
+              exchange.getRequestURI().toString(),
+              IS_LOCAL);
+
+      AuthorizationCodeResponseUrl authResponse = new AuthorizationCodeResponseUrl(requestURL);
+
+      // check for user-denied error
+      if (authResponse.getError() != null) {
+        logger.warn("Authorization DENIED: {} Redirecting to /error", authResponse.getError());
+        return redirect;
+      }
+
+      // retrieve cookie from exchange
+      Map<String, HttpCookie> httpCookies = ReferenceApiUtils.getCookies(requestHeaders);
+      HttpCookie encodedIdCookie = httpCookies.get(JsonKeys.ID_COOKIE_KEY);
       Preconditions.checkArgument(
-          !Strings.isNullOrEmpty(encodedIdCookie), "Missing encodedIdCookie");
-      // Valid job must be present
-      Preconditions.checkArgument(
-          !Strings.isNullOrEmpty(encodedIdCookie), "Encoded Id cookie required");
-      UUID jobId = ReferenceApiUtils.decodeJobId(encodedIdCookie);
+          encodedIdCookie != null && !Strings.isNullOrEmpty(encodedIdCookie.getValue()),
+          "Encoded Id cookie required");
+
+      UUID jobId = ReferenceApiUtils.decodeJobId(encodedIdCookie.getValue());
+
+      // TODO(#258): Check job ID in state token, was broken during local demo
+      // UUID jobIdFromState = ReferenceApiUtils.decodeJobId(authResponse.getState());
+
+      // // TODO: Remove sanity check
+      // Preconditions.checkState(
+      //     jobIdFromState.equals(jobId),
+      //     "Job id in cookie [%s] and request [%s] should match",
+      //     jobId,
+      //     jobIdFromState);
 
       PortabilityJob job = store.findJob(jobId);
       Preconditions.checkNotNull(job, "existing job not found for jobId: %s", jobId);
 
       // TODO: Determine service from job or from authUrl path?
       AuthMode authMode = ReferenceApiUtils.getAuthMode(exchange.getRequestHeaders());
-
-      // TODO: Determine service from job or from authUrl path?
       String service = (authMode == AuthMode.EXPORT) ? job.exportService() : job.importService();
       Preconditions.checkState(
           !Strings.isNullOrEmpty(service),
           "service not found, service: %s authMode: %s, jobId: %s",
           service,
           authMode,
-          jobId);
-
-      Preconditions.checkArgument(
-          !Strings.isNullOrEmpty(request.getUsername()), "Missing valid username");
-      Preconditions.checkArgument(
-          !Strings.isNullOrEmpty(request.getPassword()), "Missing password");
+          jobId.toString());
 
       AuthDataGenerator generator =
-          registry.getAuthDataGenerator(service, job.transferDataType(), AuthMode.EXPORT);
+          registry.getAuthDataGenerator(service, job.transferDataType(), authMode);
       Preconditions.checkNotNull(
           generator,
           "Generator not found for type: %s, service: %s",
           job.transferDataType(),
           service);
 
-      // TODO: change signature to pass UUID
-
-      // Generate and store auth data
-      AuthData authData =
-          generator.generateAuthData(
-              baseApiUrl,
-              request.getUsername(),
-              jobId.toString(),
-              null,
-              request.getPassword());
-      Preconditions.checkNotNull(authData, "Auth data should not be null");
-
       // Obtain the session key for this job
       String encodedSessionKey = job.jobAuthorization().sessionSecretKey();
       SecretKey key =
           symmetricKeyGenerator.parse(BaseEncoding.base64Url().decode(encodedSessionKey));
+
+      // Retrieve initial auth data, if it existed
+      AuthData initialAuthData = null;
+      String encryptedInitialAuthData =
+          (authMode == AuthMode.EXPORT)
+              ? job.jobAuthorization().encryptedInitialExportAuthData()
+              : job.jobAuthorization().encryptedInitialImportAuthData();
+      if (encryptedInitialAuthData != null) {
+        // Retrieve and parse the session key from the job
+        // Decrypt and deserialize the object
+        String serialized = DecrypterFactory.create(key).decrypt(encryptedInitialAuthData);
+        initialAuthData = objectMapper.readValue(serialized, AuthData.class);
+      }
+
+      // TODO: Use UUID instead of UUID.toString()
+      // Generate auth data
+      AuthData authData =
+          generator.generateAuthData(
+              baseApiUrl,
+              authResponse.getCode(),
+              jobId.toString(),
+              initialAuthData,
+              null);
+      Preconditions.checkNotNull(authData, "Auth data should not be null");
+
       // Serialize and encrypt the auth data
       String serialized = objectMapper.writeValueAsString(authData);
       String encryptedAuthData = EncrypterFactory.create(key).encrypt(serialized);
       // Set new cookie
       ReferenceApiUtils.setCookie(exchange.getResponseHeaders(), encryptedAuthData, authMode);
 
-      return new DataTransferResponse(
-          job.exportService(),
-          job.importService(),
-          job.transferDataType(),
-          Status.INPROCESS,
-          baseUrl
-              + (authMode == AuthMode.EXPORT
-                  ? FrontendConstantUrls.URL_NEXT_PAGE
-                  : FrontendConstantUrls.URL_COPY_PAGE));
-
+      redirect = baseUrl
+              + ((authMode == AuthMode.EXPORT)
+                  ? ReferenceApiUtils.FrontendConstantUrls.URL_NEXT_PAGE
+                  : ReferenceApiUtils.FrontendConstantUrls.URL_COPY_PAGE);
     } catch (Exception e) {
-      logger.debug("Exception occurred while trying to handle request: {}", e);
+      logger.error("Error handling request: {}", e);
       throw e;
     }
+
+    return redirect;
   }
 }
