+++ extensions/transport/portability-transport-jdk/src/main/java/org/dataportabilityproject/transport/jdk/http/SimpleLoginSubmitHandler.java
@@ -13,24 +13,25 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.dataportabilityproject.api.reference;
+package org.dataportabilityproject.transport.jdk.http;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.base.Preconditions;
 import com.google.common.base.Strings;
 import com.google.common.io.BaseEncoding;
-import com.google.common.net.HttpHeaders;
 import com.google.inject.Inject;
 import com.google.inject.name.Named;
 import com.sun.net.httpserver.Headers;
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
 import java.io.IOException;
-import java.util.Map;
+import java.nio.charset.StandardCharsets;
 import java.util.UUID;
 import javax.crypto.SecretKey;
+import org.apache.http.HttpHeaders;
 import org.dataportabilityproject.api.launcher.TypeManager;
-import org.dataportabilityproject.security.DecrypterFactory;
+import org.dataportabilityproject.transport.jdk.http.ReferenceApiUtils.FrontendConstantUrls;
+import org.dataportabilityproject.transport.jdk.http.ReferenceApiUtils.HttpMethods;
 import org.dataportabilityproject.security.EncrypterFactory;
 import org.dataportabilityproject.security.SymmetricKeyGenerator;
 import org.dataportabilityproject.spi.cloud.storage.JobStore;
@@ -38,20 +39,23 @@ import org.dataportabilityproject.spi.cloud.types.PortabilityJob;
 import org.dataportabilityproject.spi.api.auth.AuthDataGenerator;
 import org.dataportabilityproject.spi.api.auth.AuthServiceProviderRegistry;
 import org.dataportabilityproject.spi.api.auth.AuthServiceProviderRegistry.AuthMode;
+import org.dataportabilityproject.types.client.transfer.DataTransferResponse;
+import org.dataportabilityproject.types.client.transfer.DataTransferResponse.Status;
+import org.dataportabilityproject.types.client.transfer.SimpleLoginRequest;
 import org.dataportabilityproject.types.transfer.auth.AuthData;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- * HttpHandler for callbacks from Oauth1 authorization flow. Redirects client request to: - the next
+ * HttpHandler for SimpleLoginSubmit authorization flow. Redirects client request to: - the next
  * authorization (if this is after the source service auth) or - the copy page (if this is after the
  * destination service auth)
  */
-final class OauthCallbackHandler implements HttpHandler {
-  public static final String PATH = "/callback1/";
-  // TODO: obtain from flags
-  private static final boolean IS_LOCAL = true;
-  private final Logger logger = LoggerFactory.getLogger(OauthCallbackHandler.class);
+final class SimpleLoginSubmitHandler implements HttpHandler {
+
+  public static final String PATH = "/_/simpleLoginSubmit";
+  private static final Logger logger = LoggerFactory.getLogger(SimpleLoginSubmitHandler.class);
+
   private final AuthServiceProviderRegistry registry;
   private final JobStore store;
   private final SymmetricKeyGenerator symmetricKeyGenerator;
@@ -60,7 +64,7 @@ final class OauthCallbackHandler implements HttpHandler {
   private final String baseApiUrl;
 
   @Inject
-  OauthCallbackHandler(
+  SimpleLoginSubmitHandler(
       JobStore store,
       AuthServiceProviderRegistry registry,
       SymmetricKeyGenerator symmetricKeyGenerator,
@@ -75,116 +79,102 @@ final class OauthCallbackHandler implements HttpHandler {
     this.baseApiUrl = baseApiUrl;
   }
 
-  @Override
   public void handle(HttpExchange exchange) throws IOException {
-    // Add .* to resource path as this path will be of the form /callback1/SERVICEPROVIDER
     Preconditions.checkArgument(
-        ReferenceApiUtils.validateRequest(exchange, ReferenceApiUtils.HttpMethods.GET, PATH + ".*"));
+        ReferenceApiUtils.validateRequest(exchange, HttpMethods.POST, PATH));
     logger.debug("received request: {}", exchange.getRequestURI());
 
-    String redirect = handleExchange(exchange);
-    logger.debug("redirecting to {}", redirect);
-    exchange.getResponseHeaders().set(HttpHeaders.LOCATION, redirect);
-    exchange.sendResponseHeaders(303, -1);
+    DataTransferResponse response = handleExchange(exchange);
+
+    logger.debug("redirecting to: {}", response.getNextUrl());
+    // Mark the response as type Json and send
+    exchange
+        .getResponseHeaders()
+        .set(
+            HttpHeaders.CONTENT_TYPE, "application/json; charset=" + StandardCharsets.UTF_8.name());
+    exchange.sendResponseHeaders(200, 0);
+
+    objectMapper.writeValue(exchange.getResponseBody(), response);
   }
 
-  private String handleExchange(HttpExchange exchange) throws IOException {
-    String redirect = "/error";
+  DataTransferResponse handleExchange(HttpExchange exchange) throws IOException {
 
+    Headers requestHeaders = exchange.getRequestHeaders();
     try {
-      Headers requestHeaders = exchange.getRequestHeaders();
-
-      // Get the URL for the request - needed for the authorization.
-
-      String requestURL =
-          ReferenceApiUtils.createURL(
-              requestHeaders.getFirst(HttpHeaders.HOST),
-              exchange.getRequestURI().toString(),
-              IS_LOCAL);
-
-      Map<String, String> requestParams = ReferenceApiUtils.getRequestParams(exchange);
+      SimpleLoginRequest request =
+          objectMapper.readValue(exchange.getRequestBody(), SimpleLoginRequest.class);
 
       String encodedIdCookie = ReferenceApiUtils.getCookie(requestHeaders, JsonKeys.ID_COOKIE_KEY);
       Preconditions.checkArgument(
           !Strings.isNullOrEmpty(encodedIdCookie), "Missing encodedIdCookie");
-
-      String oauthToken = requestParams.get("oauth_token");
-      Preconditions.checkArgument(!Strings.isNullOrEmpty(oauthToken), "Missing oauth_token");
-
-      String oauthVerifier = requestParams.get("oauth_verifier");
-      Preconditions.checkArgument(!Strings.isNullOrEmpty(oauthVerifier), "Missing oauth_verifier");
-
       // Valid job must be present
       Preconditions.checkArgument(
           !Strings.isNullOrEmpty(encodedIdCookie), "Encoded Id cookie required");
       UUID jobId = ReferenceApiUtils.decodeJobId(encodedIdCookie);
 
       PortabilityJob job = store.findJob(jobId);
-
       Preconditions.checkNotNull(job, "existing job not found for jobId: %s", jobId);
 
       // TODO: Determine service from job or from authUrl path?
       AuthMode authMode = ReferenceApiUtils.getAuthMode(exchange.getRequestHeaders());
+
+      // TODO: Determine service from job or from authUrl path?
       String service = (authMode == AuthMode.EXPORT) ? job.exportService() : job.importService();
       Preconditions.checkState(
           !Strings.isNullOrEmpty(service),
           "service not found, service: %s authMode: %s, jobId: %s",
           service,
           authMode,
-          jobId.toString());
+          jobId);
+
+      Preconditions.checkArgument(
+          !Strings.isNullOrEmpty(request.getUsername()), "Missing valid username");
+      Preconditions.checkArgument(
+          !Strings.isNullOrEmpty(request.getPassword()), "Missing password");
 
       AuthDataGenerator generator =
-          registry.getAuthDataGenerator(service, job.transferDataType(), authMode);
+          registry.getAuthDataGenerator(service, job.transferDataType(), AuthMode.EXPORT);
       Preconditions.checkNotNull(
           generator,
           "Generator not found for type: %s, service: %s",
           job.transferDataType(),
           service);
 
+      // TODO: change signature to pass UUID
+
+      // Generate and store auth data
+      AuthData authData =
+          generator.generateAuthData(
+              baseApiUrl,
+              request.getUsername(),
+              jobId.toString(),
+              null,
+              request.getPassword());
+      Preconditions.checkNotNull(authData, "Auth data should not be null");
+
       // Obtain the session key for this job
       String encodedSessionKey = job.jobAuthorization().sessionSecretKey();
       SecretKey key =
           symmetricKeyGenerator.parse(BaseEncoding.base64Url().decode(encodedSessionKey));
-
-      // Retrieve initial auth data, if it existed
-      AuthData initialAuthData = null;
-      String encryptedInitialAuthData =
-          (authMode == AuthMode.EXPORT)
-              ? job.jobAuthorization().encryptedInitialExportAuthData()
-              : job.jobAuthorization().encryptedInitialImportAuthData();
-      if (encryptedInitialAuthData != null) {
-        // Retrieve and parse the session key from the job
-        // Decrypt and deserialize the object
-        String serialized = DecrypterFactory.create(key).decrypt(encryptedInitialAuthData);
-        initialAuthData = objectMapper.readValue(serialized, AuthData.class);
-      }
-
-      Preconditions.checkNotNull(
-          initialAuthData, "Initial AuthData expected during Oauth 1.0 flow");
-
-      // TODO: Use UUID instead of UUID.toString()
-      // Generate auth data
-      AuthData authData =
-          generator.generateAuthData(
-              baseApiUrl, oauthVerifier, jobId.toString(), initialAuthData, null);
-      Preconditions.checkNotNull(authData, "Auth data should not be null");
-
       // Serialize and encrypt the auth data
       String serialized = objectMapper.writeValueAsString(authData);
       String encryptedAuthData = EncrypterFactory.create(key).encrypt(serialized);
       // Set new cookie
       ReferenceApiUtils.setCookie(exchange.getResponseHeaders(), encryptedAuthData, authMode);
 
-      redirect =
+      return new DataTransferResponse(
+          job.exportService(),
+          job.importService(),
+          job.transferDataType(),
+          Status.INPROCESS,
           baseUrl
-              + ((authMode == AuthMode.EXPORT)
-                  ? ReferenceApiUtils.FrontendConstantUrls.URL_NEXT_PAGE
-                  : ReferenceApiUtils.FrontendConstantUrls.URL_COPY_PAGE);
+              + (authMode == AuthMode.EXPORT
+                  ? FrontendConstantUrls.URL_NEXT_PAGE
+                  : FrontendConstantUrls.URL_COPY_PAGE));
+
     } catch (Exception e) {
-      logger.error("Error handling request", e);
+      logger.debug("Exception occurred while trying to handle request: {}", e);
       throw e;
     }
-
-    return redirect;
   }
 }
