+++ extensions/transport/portability-transport-jdk/src/main/resources/META-INF/services/org.dataportabilityproject.spi.service.extension.ServiceExtension
@@ -1,0 +1,1 @@
+org.dataportabilityproject.transport.jdk.JdkTransportExtension
