+++ portability-api/src/main/java/org/dataportabilityproject/api/ApiExtensionContext.java
@@ -16,14 +16,16 @@
 package org.dataportabilityproject.api;
 
 import com.google.common.base.Preconditions;
-import java.util.HashMap;
-import java.util.Map;
+import com.google.inject.Injector;
 import org.dataportabilityproject.api.launcher.Constants.Environment;
 import org.dataportabilityproject.api.launcher.ExtensionContext;
+import org.dataportabilityproject.api.launcher.Flag;
 import org.dataportabilityproject.api.launcher.Logger;
 import org.dataportabilityproject.api.launcher.TypeManager;
-import org.dataportabilityproject.api.launcher.Flag;
 import org.dataportabilityproject.config.extension.SettingsExtension;
+
+import java.util.HashMap;
+import java.util.Map;
 
 /**
  * Provides a context for initializing extensions.
@@ -34,6 +36,7 @@ public class ApiExtensionContext implements ExtensionContext {
   private final Map<Class<?>, Object> registered = new HashMap<>();
   private final TypeManager typeManager;
   private final SettingsExtension settingsExtension;
+  private Injector injector;
 
   // Required settings
   private final String cloud;
