+++ portability-bootstrap-vm/src/main/java/org/dataportabilityproject/bootstrap/vm/SingleVMMain.java
@@ -2,7 +2,6 @@ package org.dataportabilityproject.bootstrap.vm;
 
 import com.google.common.util.concurrent.UncaughtExceptionHandlers;
 import org.dataportabilityproject.api.ApiMain;
-import org.dataportabilityproject.api.reference.ReferenceApiServer;
 import org.dataportabilityproject.transfer.WorkerMain;
 
 import javax.net.ssl.KeyManagerFactory;
@@ -42,9 +41,7 @@ public class SingleVMMain {
     ApiMain apiMain = new ApiMain();
 
     try (InputStream stream =
-        ReferenceApiServer.class
-            .getClassLoader()
-            .getResourceAsStream("demo-selfsigned-keystore.jks")) {
+        ApiMain.class.getClassLoader().getResourceAsStream("demo-selfsigned-keystore.jks")) {
       if (stream == null) {
         throw new IllegalArgumentException("Demo keystore was not found");
       }
