+++ settings.gradle
@@ -22,7 +22,7 @@ include ':portability-api', ':portability-transfer', ':portability-bootstrap-vm'
 include ':extensions:cloud:portability-cloud-local', ':extensions:cloud:portability-cloud-google', ':extensions:cloud:portability-cloud-microsoft'
 
 // Transport Extensions
-include ':extensions:transport:portability-transport-jettyrest'
+include ':extensions:transport:portability-transport-jdk',':extensions:transport:portability-transport-jettyrest'
 
 // Flickr Auth and Transfer Extensions
 include ':extensions:auth:portability-auth-flickr'
